+++ bauwerk/benchmarks.py
@@ -47,23 +47,6 @@ class ContParamDist(ParamDist):
         return self.fn(low=self.low, high=self.high)
 
 
-@dataclass
-class CfgDist:
-    """Distribution over env configurations.
-
-    Made up of ParamDist's.
-    """
-
-    variable_params: Dict[str, ParamDist]
-    fixed_params: Optional[Dict] = None
-
-    def sample(self):
-        variable_params_dict = dict(
-            (name, dist.sample()) for name, dist in self.variable_params.items()
-        )
-        return bauwerk.EnvConfig(**variable_params_dict, **self.fixed_params)
-
-
 def sample_cfg_dist(self) -> bauwerk.EnvConfig:
     """Sample from CfgDist."""
 
@@ -153,10 +136,14 @@ class BuildDist(Benchmark):
         """Building distribution.
 
         Args:
-            cfg_dist (CfgDist): distribution over bauwerk env configs.
-            seed (Optional[int], optional): Random seed. Defaults to None.
-            num_train_tasks (int, optional): Number of training tasks. Defaults to 20.
-            num_test_tasks (int, optional): Number of test tasks. Defaults to 10.
+            cfg_dist (CfgDist): distribution over bauwerk
+                env configs.
+            seed (int, optional): Random seed.
+                Defaults to None.
+            num_train_tasks (int, optional): Number of training tasks.
+                Defaults to 20.
+            num_test_tasks (int, optional): Number of test tasks.
+                Defaults to 10.
             episode_len: (int, optional): Length of episode in distribution
                 environments. If not set, defaults to distribution configuration.
             dtype (Union[str, np.dtype], optional): data type to be returned and
@@ -165,6 +152,7 @@ class BuildDist(Benchmark):
             env_kwargs (dict, optional): parameters to pass when creating environment.
                 This should not be used when evaluating on pre-defined benchmark.
                 Defaults to None.
+
         """
         super().__init__()
 
@@ -239,11 +227,28 @@ class BuildDist(Benchmark):
         return env
 
 
-class BuildDistB(BuildDist):
-    """Bauwerk building distribution B: varying battery sizes."""
+class BuildDistA(BuildDist):
+    """Bauwerk building distribution A: identical houses, no variation."""
 
     def __init__(self, **kwargs):
-        """Bauwerk building distribution B: varying battery sizes."""
+        """Bauwerk building distribution A:
+
+        Identical houses, no variation."""
+        cfg_dist = CfgDist(
+            battery_size=7.5,
+            episode_len=24 * 30,
+            grid_peak_threshold=2.0,
+        )
+        super().__init__(**kwargs, cfg_dist=cfg_dist)
+
+
+class BuildDistB(BuildDist):
+    """Bauwerk building distribution B:"""
+
+    def __init__(self, **kwargs):
+        """Bauwerk building distribution B:
+
+        Houses with varying battery size (0.5kWh to 20kWh)."""
         cfg_dist = CfgDist(
             battery_size=ContParamDist(
                 low=0.5,
@@ -251,5 +256,95 @@ class BuildDistB(BuildDist):
                 fn=np.random.uniform,
             ),
             episode_len=24 * 30,
+            grid_peak_threshold=2.0,
+        )
+        super().__init__(**kwargs, cfg_dist=cfg_dist)
+
+
+class BuildDistC(BuildDist):
+    """Bauwerk building distribution C:"""
+
+    def __init__(self, **kwargs):
+        """Bauwerk building distribution C.
+
+        Houses with varying solar (multiplier: 0.5 to 5) and
+        battery sizes (0.5 to 20kWh)."""
+        cfg_dist = CfgDist(
+            battery_size=ContParamDist(
+                low=0.5,
+                high=20,
+                fn=np.random.uniform,
+            ),
+            solar_scaling_factor=ContParamDist(
+                low=0.5,
+                high=5,
+                fn=np.random.uniform,
+            ),
+            episode_len=24 * 30,
+            grid_peak_threshold=2.0,
+        )
+        super().__init__(**kwargs, cfg_dist=cfg_dist)
+
+
+class BuildDistD(BuildDist):
+    """Bauwerk building distribution D: varying battery, load and solar sizes."""
+
+    def __init__(self, **kwargs):
+        """Bauwerk building distribution D.
+
+        Houses with varying battery (0.5 to 20kWh),
+        load (multiplier: 0.5 to 5) and solar sizes
+        (multiplier: 0.5 to 5).
+        This distribution is effectively like differently sized houses."""
+        cfg_dist = CfgDist(
+            battery_size=ContParamDist(
+                low=0.5,
+                high=20,
+                fn=np.random.uniform,
+            ),
+            solar_scaling_factor=ContParamDist(
+                low=0.5,
+                high=5,
+                fn=np.random.uniform,
+            ),
+            load_scaling_factor=ContParamDist(
+                low=0.5,
+                high=5,
+                fn=np.random.uniform,
+            ),
+            episode_len=24 * 30,
+            grid_peak_threshold=2.0,
+        )
+        super().__init__(**kwargs, cfg_dist=cfg_dist)
+
+
+class BuildDistE(BuildDist):
+    """Bauwerk building distribution E."""
+
+    def __init__(self, **kwargs):
+        """Bauwerk building distribution E.
+
+        Same as Bauwerk building distribution D,
+        other than adding irreducible noise."""
+        cfg_dist = CfgDist(
+            battery_size=ContParamDist(
+                low=0.5,
+                high=20,
+                fn=np.random.uniform,
+            ),
+            solar_scaling_factor=ContParamDist(
+                low=0.5,
+                high=5,
+                fn=np.random.uniform,
+            ),
+            load_scaling_factor=ContParamDist(
+                low=0.5,
+                high=5,
+                fn=np.random.uniform,
+            ),
+            load_noise_magnitude=2.0,
+            solar_noise_magnitude=2.0,
+            episode_len=24 * 30,
+            grid_peak_threshold=2.0,
         )
         super().__init__(**kwargs, cfg_dist=cfg_dist)